Rename LcovMerger to CoverageOutputGenerator.
