+++ tools/BUILD
@@ -26,8 +26,8 @@ filegroup(
         "//tools/osx:srcs",
         "//tools/osx/crosstool:srcs",
         "//tools/test:srcs",
-        "//tools/test/LcovMerger/java/com/google/devtools/lcovmerger:srcs",
-        "//tools/test/LcovMerger/javatests/com/google/devtools/lcovmerger:srcs",
+        "//tools/test/CoverageOutputGenerator/java/com/google/devtools/coverageoutputgenerator:srcs",
+        "//tools/test/CoverageOutputGenerator/javatests/com/google/devtools/coverageoutputgenerator:srcs",
         "//tools/python:srcs",
         "//tools/runfiles:srcs",
         "//tools/sh:srcs",
@@ -59,7 +59,7 @@ filegroup(
         "//tools/python:embedded_tools",
         "//tools/runfiles:embedded_tools",
         "//tools/test:embedded_tools",
-        "//tools/test/LcovMerger/java/com/google/devtools/lcovmerger:embedded_tools",
+        "//tools/test/CoverageOutputGenerator/java/com/google/devtools/coverageoutputgenerator:embedded_tools",
         "//tools/osx/crosstool:srcs",
         "//tools/osx:srcs",
         "//tools/sh:embedded_tools",
