+++ tools/test/CoverageOutputGenerator/java/com/google/devtools/coverageoutputgenerator/BUILD.tools
@@ -1,0 +1,13 @@
+package(default_visibility = ["//visibility:public"])
+
+java_import(
+  name = "all_lcov_merger_lib",
+  jars = ["all_lcov_merger_tools_deploy.jar"],
+)
+
+java_binary(
+  name = "Main",
+  create_executable = 1,
+  main_class = "com.google.devtools.coverageoutputgenerator.Main",
+  runtime_deps = [":all_lcov_merger_lib"],
+)
