--- tools/test/LcovMerger/java/com/google/devtools/lcovmerger/BUILD.tools
@@ -1,13 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-java_import(
-  name = "all_lcov_merger_lib",
-  jars = ["all_lcov_merger_tools_deploy.jar"],
-)
-
-java_binary(
-  name = "Main",
-  create_executable = 1,
-  main_class = "com.google.devtools.lcovmerger.Main",
-  runtime_deps = [":all_lcov_merger_lib"],
-)
