--- tools/test/LcovMerger/java/com/google/devtools/lcovmerger/Parser.java
@@ -1,23 +1,0 @@
-// Copyright 2018 The Bazel Authors. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//    http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package com.google.devtools.lcovmerger;
-
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.List;
-
-interface Parser {
-  List<SourceFileCoverage> parse(InputStream inputStream) throws IOException;
-}
